P2T4

+++ pointers_arrays_strings/4-rev_array.c
@@ -6,7 +6,7 @@
 */
 void reverse_array(int *a, int n)
 {
-int i, temp;
+long int i, temp;
 while (a[i] != '\0')
 {
 i++;
